rename to compact
