+++ src/query/storages/fuse/fuse/src/operations/mutation/compact/compact_source.rs
