+++ src/query/storages/fuse/fuse/src/operations/mutation/compact/segment_compact_mutator.rs
