+++ src/query/storages/fuse/fuse/src/operations/mutation/mod.rs
@@ -15,7 +15,7 @@
 pub mod abort_operation;
 pub mod base_mutator;
 pub mod block_filter;
-mod compact_mutation;
+mod compact;
 pub mod deletion_mutator;
 pub mod recluster_mutator;
 
@@ -23,12 +23,12 @@ pub use abort_operation::AbortOperation;
 pub use base_mutator::BaseMutator;
 pub use block_filter::all_the_columns_ids;
 pub use block_filter::delete_from_block;
-pub use compact_mutation::BlockCompactMutator;
-pub use compact_mutation::CompactSink;
-pub use compact_mutation::CompactSource;
-pub use compact_mutation::CompactTransform;
-pub use compact_mutation::SegmentCompactMutator;
-pub use compact_mutation::SegmentCompactionState;
-pub use compact_mutation::SegmentCompactor;
+pub use compact::BlockCompactMutator;
+pub use compact::CompactSink;
+pub use compact::CompactSource;
+pub use compact::CompactTransform;
+pub use compact::SegmentCompactMutator;
+pub use compact::SegmentCompactionState;
+pub use compact::SegmentCompactor;
 pub use deletion_mutator::DeletionMutator;
 pub use recluster_mutator::ReclusterMutator;
